Clean up query manager, store processed initial & virtual columns in it
