+++ mathesar_ui/src/systems/query-builder/QueryModel.ts
@@ -2,7 +2,6 @@ import type {
   QueryInstanceInitialColumn,
   QueryInstanceTransformation,
 } from '@mathesar/api/queries/queryList';
-import { isDefinedNonNullable } from '@mathesar-component-library';
 import type { UnsavedQueryInstance } from '@mathesar/stores/queries';
 import QueryFilterTransformationModel from './QueryFilterTransformationModel';
 import QuerySummarizationTransformationModel from './QuerySummarizationTransformationModel';
