+++ mathesar_ui/src/systems/query-builder/column-selection-pane/SelectableColumnTree.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import type { ColumnWithLink } from '../QueryManager';
+  import type { ColumnWithLink } from '../utils';
   import SelectableColumn from './SelectableColumn.svelte';
   import TableGroupCollapsible from './TableGroupCollapsible.svelte';
